move components related to forms to a Forms folder
