+++ resources/js/Components/Forms/TextAreaInput.vue
